Extract protocol event schemas
